--- src/main/java/com/sopkaton/web2/common/MissionNotFoundException.java
@@ -1,4 +1,0 @@
-package com.sopkaton.web2.common;
-
-public class MissionNotFoundException extends CoreException {
-}
