--- src/main/java/com/sopkaton/web2/common/TeamNotFoundException.java
@@ -1,4 +1,0 @@
-package com.sopkaton.web2.common;
-
-public class TeamNotFoundException extends CoreException {
-}
