+++ src/main/java/com/sopkaton/web2/common/api/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.sopkaton.web2.common;
+package com.sopkaton.web2.common.api;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +11,7 @@ public enum ErrorCode {
     USER_NOT_FOUND(40401),
     MISSION_NOT_FOUND(40402),
     USER_TEAM_NOT_FOUND(40403),
-    TEAM_NOT_FOUND(40400);
+    TEAM_NOT_FOUND(40400), INTERNAL_SERVER_ERROR(50000);
 
     private final int customErrorCode;
 }
