+++ src/main/java/com/sopkaton/web2/common/api/SuccessCode.java
@@ -1,4 +1,4 @@
-package com.sopkaton.web2.common;
+package com.sopkaton.web2.common.api;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
