+++ src/main/java/com/sopkaton/web2/common/core/CoreException.java
@@ -1,4 +1,4 @@
-package com.sopkaton.web2.common;
+package com.sopkaton.web2.common.core;
 
 public class CoreException extends RuntimeException {
 }
