+++ src/main/java/com/sopkaton/web2/common/core/MemberNotFoundException.java
@@ -1,0 +1,6 @@
+package com.sopkaton.web2.common.core;
+
+import com.sopkaton.web2.common.core.CoreException;
+
+public class MemberNotFoundException extends CoreException {
+}
