+++ src/main/java/com/sopkaton/web2/common/core/UserTeamNotFoundException.java
@@ -1,0 +1,6 @@
+package com.sopkaton.web2.common.core;
+
+import com.sopkaton.web2.common.core.CoreException;
+
+public class UserTeamNotFoundException extends CoreException {
+}
