--- src/main/java/com/sopkaton/web2/controller/ExceptionHandler.java
@@ -1,9 +1,0 @@
-package com.sopkaton.web2.controller;
-
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ExceptionHandler {
-
-}
