+++ src/main/java/com/sopkaton/web2/controller/GlobalExceptionHandler.java
@@ -1,0 +1,25 @@
+package com.sopkaton.web2.controller;
+
+
+import com.sopkaton.web2.common.api.CustomException;
+import com.sopkaton.web2.common.api.ErrorCode;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(CustomException.class)
+    public ApiResponse<Void> handleCustomException(CustomException e) {
+        return ApiResponse.fail(e.getErrorCode());
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ApiResponse<Void>> handleException(Exception e) {
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                .body(ApiResponse.fail(ErrorCode.INTERNAL_SERVER_ERROR));
+    }
+}
+
