+++ src/main/java/com/sopkaton/web2/repository/mission/MissionRepository.java
@@ -1,4 +1,4 @@
-package com.sopkaton.web2.repository;
+package com.sopkaton.web2.repository.mission;
 
 import org.springframework.data.jpa.repository.JpaRepository;
