+++ src/main/java/com/sopkaton/web2/repository/team/TeamRepository.java
@@ -1,4 +1,4 @@
-package com.sopkaton.web2.repository;
+package com.sopkaton.web2.repository.team;
 
 import org.springframework.data.jpa.repository.JpaRepository;
