+++ src/main/java/com/sopkaton/web2/repository/userteam/UserTeamRepository.java
@@ -1,5 +1,7 @@
-package com.sopkaton.web2.repository;
+package com.sopkaton.web2.repository.userteam;
 
+import com.sopkaton.web2.repository.team.Team;
+import com.sopkaton.web2.repository.user.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
